# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenAPI",
        "description": "L5 Swagger OpenAPI description",
        "contact": {
            "email": "darius@matulion.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký người dùng mới",
                "description": "\n     *      - Endpoint này cho phép đăng ký người dùng mới vào hệ thống.\n     *      - Trả về thông tin của người dùng đã đăng ký.\n     *      - Role được sử dụng là nhân viên và sinh viên",
                "operationId": "ab15e8b534800db0828c4efe3a6099fa",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "role": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "John Doe",
                                                "email": "john.doe@example.com",
                                                "phone": "123456789",
                                                "role": "user"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng ký người dùng thành công"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi máy chủ nội bộ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "tên không thể để trống"
                                                ],
                                                "email": [
                                                    "email không thể để trống"
                                                ]
                                            }
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập người dùng",
                "description": "\n     *      - Endpoint này cho phép người dùng đăng nhập vào hệ thống.\n     *      - Trả về thông tin của người dùng đã đăng nhập, bao gồm cả token đăng nhập.\n     *      - Role được sử dụng là cả ba role nhân viên ,quản lí ,sinh viên",
                "operationId": "306b0884672cb3eed691231729c47fbb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john.doe@example.com"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "123456789"
                                                },
                                                "role": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "api-token"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Đăng nhập người dùng thành công"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi máy chủ nội bộ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thông tin đăng nhập không chính xác. Vui lòng thử lại"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 500
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}